use after_flush_postexec event instead of after_flush
Ran into a weird exception with another project that I didn't really
understand.  The error was:

Deferred loader for attribute 'createdts' failed to populate correctly

However, I have the exact same column type & setup in the SAV tests
and it ran fine, so not really sure what to do with that.